RAG_Q&A_Chatbot/requirements.txt
ipykernel
langchain-community
python-dotenv
streamlit
langchain-chroma
langchain-core
docx2txt
